chore: move Breadcrumb component impls into spi/
Fixes #4443
